requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.2.3
django-environ==0.11.1
djangorestframework==3.14.0
drf-yasg==1.21.7
exceptiongroup==1.1.3
factory-boy==3.3.0
Faker==19.3.1
flake8==6.1.0
freezegun==1.2.2
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
mccabe==0.7.0
openapi-codec==1.3.2
packaging==23.1
pluggy==1.3.0
psycopg2-binary==2.9.6
pycodestyle==2.11.0
pyflakes==3.1.0
pytest==7.4.0
pytest-django==4.5.2
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
simplejson==3.19.1
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==2.0.4
